LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/modal-dialog-selection-expected.txt

123456789A

PASS By default, text inside a modal dialog can be selected
PASS Clicking on backdrop doesn't select text
FAIL 'user-select: none' prevents text from being selected assert_equals: expected "" but got "345678"
